+++ src/hooks/index.ts
@@ -1,15 +1,16 @@
 // Rainy Cowork - Hooks Barrel Export
 
-export { useTauriTask } from './useTauriTask';
-export { useAIProvider } from './useAIProvider';
-export { useStreaming } from './useStreaming';
-export { useUsageAnalytics } from './useUsageAnalytics';
-export { useCoworkStatus } from './useCoworkStatus';
-export { useAgent } from './useAgent';
-export { useFolderManager } from './useFolderManager';
-export { useCoworkKeys } from './useCoworkKeys';
-export { useCoworkBilling } from './useCoworkBilling';
-export { useTheme } from './useTheme';
-export { useWorkspace } from './useWorkspace';
-export { useFileVersioning } from './useFileVersioning';
-export { useIntelligentRouter } from './useIntelligentRouter';
+export { useTauriTask } from "./useTauriTask";
+export { useAIProvider } from "./useAIProvider";
+export { useStreaming } from "./useStreaming";
+export { useUsageAnalytics } from "./useUsageAnalytics";
+export { useCoworkStatus } from "./useCoworkStatus";
+export { useAgent } from "./useAgent";
+export { useFolderManager } from "./useFolderManager";
+export { useCoworkKeys } from "./useCoworkKeys";
+export { useCoworkBilling } from "./useCoworkBilling";
+export { useTheme } from "./useTheme";
+export { useWorkspace } from "./useWorkspace";
+export { useFileVersioning } from "./useFileVersioning";
+export { useIntelligentRouter } from "./useIntelligentRouter";
+export { useNeuralService } from "./useNeuralService";